Display territory index in middle of region
Rough approximation of where the unit count will go. Will need to use
the number mask PNG file to get a more accurate & human friendly
location.

The plan (for the short term at least) is to use this index as a sort of
ID with which we can refer to in code for things such as territory
connections.